refact(order): 주문 리팩토링

+++ order-service/src/main/java/com/rushcrew/order_service/application/command/dto/command/CreateOrderCommand.java
@@ -1,6 +1,5 @@
 package com.rushcrew.order_service.application.command.dto.command;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
@@ -23,6 +22,6 @@ public record CreateOrderCommand(
 	@Builder
 	public record OrderItemCommand(
 		UUID timeDealStockId,
-		Integer quantity
+		Long quantity
 	) {}
 }
